Moving more things out of GL into VideoCommon...
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4187 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoOGL/Src/DLCache.cpp
@@ -494,7 +494,7 @@ bool HandleDisplayList(u32 address, u32 size)
 	u64 dl_id = DLCache::CreateMapId(address, size);
 	DLCache::DLMap::iterator iter = DLCache::dl_map.find(dl_id);
 
-	stats.numDListsAlive = DLCache::dl_map.size();
+	stats.numDListsAlive = (int)DLCache::dl_map.size();
 	if (iter != DLCache::dl_map.end())
 	{
 		DLCache::CachedDisplayList &dl = iter->second;
